+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/spells/impl/AuraSpells.java
@@ -10,7 +10,7 @@ import com.robertx22.age_of_exile.database.data.stats.types.offense.Accuracy;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.AttackSpeed;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.crit.CriticalHit;
 import com.robertx22.age_of_exile.database.data.stats.types.resources.health.Health;
-import com.robertx22.age_of_exile.database.data.stats.types.resources.magic_shield.MagicShield;
+import com.robertx22.age_of_exile.database.data.stats.types.resources.magic_shield.MagicShieldRegen;
 import com.robertx22.age_of_exile.database.data.stats.types.spell_calc.CastSpeed;
 import com.robertx22.age_of_exile.database.registry.ISlashRegistryInit;
 import com.robertx22.age_of_exile.uncommon.enumclasses.Elements;
@@ -42,9 +42,9 @@ public class AuraSpells implements ISlashRegistryInit {
             )))
             .build();
 
-        SpellBuilder.aura("ms_aura", "Magic Shield Aura", new AuraSpellData(0.2F,
+        SpellBuilder.aura("ms_reg_aura", "Magic Shield Regeneration Aura", new AuraSpellData(0.1F,
             Arrays.asList(
-                new StatModifier(10, 25, MagicShield.getInstance())
+                new StatModifier(2, 5, MagicShieldRegen.getInstance())
             )))
             .build();