+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/unique_gears/UniqueGearReg.java
@@ -3,6 +3,7 @@ package com.robertx22.age_of_exile.aoe_data.database.unique_gears;
 import com.robertx22.age_of_exile.aoe_data.database.unique_gears.registrators.UniqueArmors;
 import com.robertx22.age_of_exile.aoe_data.database.unique_gears.registrators.UniqueJewelry;
 import com.robertx22.age_of_exile.aoe_data.database.unique_gears.registrators.UniqueWeapons;
+import com.robertx22.age_of_exile.aoe_data.database.unique_gears.registrators.monster_uniques.MonsterUniques;
 import com.robertx22.age_of_exile.database.registry.ISlashRegistryInit;
 
 public class UniqueGearReg implements ISlashRegistryInit {
@@ -12,5 +13,7 @@ public class UniqueGearReg implements ISlashRegistryInit {
         new UniqueArmors().registerAll();
         new UniqueJewelry().registerAll();
         new UniqueWeapons().registerAll();
+
+        new MonsterUniques().registerAll();
     }
 }