+++ src/main/java/com/robertx22/age_of_exile/mmorpg/registers/common/items/UniqueGearItemRegister.java
@@ -49,4 +49,6 @@ public class UniqueGearItemRegister extends BaseItemRegistrator {
     public Item OBSI_MIGHT_AXE = item(new BaseUniqueAxe("Axe"), "uniques/weapon/obsi_might");
     public Item JUDGEMENT_AXE = item(new BaseUniqueAxe("Axe"), "uniques/weapon/judgement");
 
+    public Item BONECHILL_RING_MI = item(new BaseUniqueRing("Ring"), "uniques/jewelry/bonechill_ring");
+
 }