+++ src/main/java/com/robertx22/age_of_exile/saveclasses/item_classes/GearTooltipUtils.java
@@ -10,6 +10,7 @@ import com.robertx22.age_of_exile.saveclasses.ExactStatData;
 import com.robertx22.age_of_exile.saveclasses.gearitem.gear_bases.IGearPartTooltip;
 import com.robertx22.age_of_exile.saveclasses.gearitem.gear_bases.TooltipInfo;
 import com.robertx22.age_of_exile.saveclasses.item_classes.tooltips.MergedStats;
+import com.robertx22.age_of_exile.uncommon.interfaces.data_items.IRarity;
 import com.robertx22.age_of_exile.uncommon.localization.Words;
 import com.robertx22.age_of_exile.uncommon.utilityclasses.TooltipUtils;
 import com.robertx22.age_of_exile.uncommon.wrappers.SText;
@@ -113,7 +114,9 @@ public class GearTooltipUtils {
             }
         }
 
-        if (gear.is_unique) {
+        if (gear.getRarity()
+            .GUID()
+            .equals(IRarity.UNIQUE_ID)) {
             if (!info.shouldShowDescriptions()) {
                 UniqueGear unique = gear.uniqueStats.getUnique(gear);
                 TooltipUtils.addUniqueDesc(tip, unique, gear);